proprietary: Pin the yet again modified CustomizationSelector
Issue: The APK would crash due to the use of the getGroupIdLevel2() fuction in telephony/java/android/telephony/TelephonyManager.java. This function is neither present in AOSP nor in LOS and appears to be a Sony-specific modification. The crash would occur in CommonUtil:isMandatorySimParmsAvailable() and SimConfigID:extractSimInfo().

Solution: Modify the CommonUtil.class and SimConfigId.class to replace the getGroupIdLevel2() function calls with getGroupIdLevel1(). As a result, the GID1 value will also be stored in GID2. However, no relevant entries exist for the CustomizationSelector that do have GID2 values. Those few were therefore also removed from the added overlays to avoid any potential issues.

Note: Two changes exist on AOSP gerrit (462638 and 462657) which would add these functions to TelephonyManager. They were abandoned, so it does not appear a relevant change is going to make it upstream any time soon.